+++ src/raysphere.cpp
@@ -3,7 +3,7 @@
 #include "math.h"
 #include "utils.h"
 
-RaySphere::RaySphere(const Vector3D& center, float radius, const RayMaterial& material)
+RaySphere::RaySphere(const Vector3D& center, double radius, const RayMaterial& material)
     : m_center(center), m_radius(radius), m_material(material)
 {
 }
@@ -14,17 +14,17 @@ bool RaySphere::intersects(const Vector3D& start, const Vector3D& direction, std
     // Line = x(t) = o + t*v
     // Sphere = (x - c)(x - c) = r^2
 
-    float A = direction.dotProduct(direction);
-    float B = 2*(start - m_center).dotProduct(direction);
-    float C = (start - m_center).dotProduct(start - m_center) - m_radius*m_radius;
-    float root = B*B - 4*A*C;
+    double A = direction.dotProduct(direction);
+    double B = 2*(start - m_center).dotProduct(direction);
+    double C = (start - m_center).dotProduct(start - m_center) - m_radius*m_radius;
+    double root = B*B - 4*A*C;
     if(root < 0)
         return false;
 
     // Thw two solutions
-    float t1 = (-B + sqrt(root))/(2*A);
-    float t2 = (-B - sqrt(root))/(2*A);
-    float ts[] = {t1, t2};
+    double t1 = (-B + sqrt(root))/(2*A);
+    double t2 = (-B - sqrt(root))/(2*A);
+    double ts[] = {t1, t2};
     for(int i=0;i<2;++i)
     {
         if(ts[i] < 0)
@@ -67,7 +67,7 @@ void RaySphere::setCenter(const Vector3D& center)
     m_center = center;
 }
 
-void RaySphere::setRadius(float radius)
+void RaySphere::setRadius(double radius)
 {
     m_radius = radius;
 }
